Move Value to Core
