+++ src/core.rs
@@ -3,4 +3,5 @@ pub mod perlin;
 pub mod perlin_surflet;
 pub mod simplex;
 pub mod super_simplex;
+pub mod value;
 pub mod worley;
